requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.0
apify_client==1.8.1
apify_shared==1.1.2
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
billiard==4.2.1
boto3==1.35.24
botocore==1.35.24
bytewax==0.21.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cog==0.12.0
coloredlogs==14.0
contourpy==1.3.0
croniter==3.0.3
cryptography==43.0.1
cycler==0.12.1
dagster==1.7.4
dagster-aws==0.23.4
dagster-celery==0.23.4
dagster-celery-k8s==0.23.4
dagster-cloud==1.7.4
dagster-cloud-cli==1.7.4
dagster-k8s==0.23.4
dagster-pipes==1.7.4
dagster-postgres==0.23.4
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
duckdb==0.9.2
durationpy==0.7
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.16.1
flower==2.0.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
github3.py==4.0.1
google-auth==2.35.0
gql==3.5.0
graphql-core==3.2.4
grpcio==1.66.1
grpcio-health-checking==1.62.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
humanize==4.10.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
kiwisolver==1.4.7
kombu==5.4.2
kubernetes==31.0.0
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.47.0
packaging==24.1
pandas==2.2.3
patito==0.7.0
pendulum==3.0.0
pex==2.20.1
pillow==10.4.0
pluggy==1.5.0
polars==1.8.0
polars-hash==0.4.8
prometheus_client==0.21.0
prompt-toolkit==3.0.36
protobuf==4.25.5
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
pytest==8.3.3
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
questionary==2.0.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rsa==4.9
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlescapy==1.0.1
starlette==0.37.2
structlog==24.4.0
tabulate==0.9.0
termcolor==2.4.0
time-machine==2.15.0
tokenizers==0.20.0
tomli==2.0.1
toposort==1.10
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchdog==5.0.2
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
xmltodict==0.13.0
yarl==1.11.1
zipp==3.20.2
